Controller prova de consistência
